Add robust functionality to handle the combination of statements under locking and normal statements. For example, database_execute_command(...); database_execute_command(...); database_execute_command(...); ... locking(...); dec_already_locked(...); dec_already_locked(...); dec_already_locked(...); .. unlock(..); ..
database_execute_command(...);
database_execute_command(...);
database_execute_command(...);
..
After all, we are not going to use the database_execute_command but
just for now since we were not able to finish to apply locking to all
database_execute_commands, we still need to mix them.
If all dec's are locked, then this modification is not needed.
This modification is fine even in the case of only locking dec's.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@2620 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 5d9c63f77e7b1a2287330203cb9ecb923212ffd3 [formerly 2cc453b87666912765bffc230dc50d3b3f3545e3] [formerly 142f0f039459d49899a3bbabe57f043d35d9e471 [formerly 39be5148e91827736722dcb99bf78fb2082005b0]]
Former-commit-id: 634139e9750a2bb18d548e7babad5367dc36f6ba [formerly d16fb3e86dcf7011c94d3d9b1880f303f30fa66c]
Former-commit-id: 402514e746b4fa0ccdd8e4f46b40dd35ab1b926b