Yume is a small tool for integrating the wellknown package management tool yum (yellowdorg updater modified) into OSCAR. It simplifies the setup and management of heterogeneous clusters, i.e. clusters running multiple distributions or even multiple architectures controlled by one single master node.
This tool is very useful for managing clusters which are running distributions
using the rpm package format. It can:
 - prepare an rpm repository (i.e. generate a metadata cache in createrepo
   format, which is used also for apt, up2date and other package managers)
 - export it through apache, /tftpboot/rpm will be visible from cluster nodes
   as http://oscar_server/repo_rpm, including the metadata cache)
 - execute yum commands applying only to this repository (locally)
 - execute yum commands on the cluster nodes applying only to this repository.
This makes installing packages, creating cluster node images, updating
revisions much simpler than with rpm.

yume-opkg is a simple ODA query tool which facilitates installation and
deinstallation of opkgs into non-native images. To be able to fully exploit
it all packages will need to be converted to generic-setup format.

A tested yum package is included in this opkg. It is called yum-oscar and
should be used in distros which do not come with their own yum (like RHEL
and mandrake). They provide the "yum" capability, and that is what yume
requires. So distros which include yum will not install it. I had no
opportunity to build yum-oscar on fedora and mandrake, so I'm not sure they
build cleanly there. yum-oscar is depending on the python version, therefore
is distro (version) dependent, but not architecture dependent.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@3911 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 7309e35640407ab7ee587a38b410a4bb92c5abfa [formerly 41de1ddce47f9e656abc2b7e3af35f3d9b3c32a7] [formerly deacd14c86b3db97c17704ea1c08c63c46a38e42 [formerly c80e7e35138d3962c45a906f1289de52268b1c5a]] [formerly 65c7b0f16770d440e67eb57f9524112ddbfa60e8 [formerly ba9d358ffe7fd1dac2a0b6df49cfe4f21efda58c] [formerly e90f32cce41bf7f31bb3a51fcbd2c4abbb2e31ab [formerly 803a91bb1288c36502fc6c28c3afc1ffff76b08c]]]
Former-commit-id: d8d2cdcd8087ed815793dee3080ab2c9c8a8db01 [formerly 06381817777afd17e73c97d73d2a7a951a305d47] [formerly 7a5268bf45155ad02c961c0bbf47dbcdcceb1fec [formerly f3ffcdee04c13c99136f071d35918620905678a7]]
Former-commit-id: faf36c32f01b1e0910312a554993c01b042e81fd [formerly cdfd33cd91296842c33c73e2e8745eb394ede67f]
Former-commit-id: 1de5b540cb3a83b9320c291d3556a3ca4528565c