This script helps integrating images created outside the OSCAR GUI into the OSCAR/SIS infrastructure. For example: images created by yume. Usage: integrate_image --name testimg --arch ia64 --disktab /opt/oscar/oscarsamples/mydisktab
The image directory testimg must contain a valid image and be located inside
/var/lib/systemimager/images.

The default architecture attributed to the image is that of the master node. The
example explicitely selects ia64.

A disktable file should be inserted into the image by using the --disktab
command line option. This generates the /etc/systemimager/autoinstall.conf file
inside the image. This is not necessary if one has copied the image from another
OSCAR cluster.

A record is created in the SIS image database. No ODA entry is created, as that one
isn't used, yet. I think.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@3912 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 26b89a26fe82e9ab8f825af0736f4009bb53a824 [formerly ec0b04fc8fc98c3000f8b8fefe10fdf83a91fae0] [formerly d33fe752445a440fa40ff5e6a2c5fefe72356635 [formerly 545ca357fd6414af7a77b9860deb407ed9ad4679]] [formerly 880148dbcdef03d808e9bec12e7e6126eeea6ae7 [formerly f312ab9f4ba5962295f972408a12fbcbd605b370] [formerly 9b6b70f2ef8ecfec9d110d5ea91aa2894c99b10d [formerly cdcdaecd87f84b580efb612f729899e09612c53a]]]
Former-commit-id: f138ba42748daa094bc00afea1f4ff5f8e55f19a [formerly fd366c10fc64c0aa56b4c6f84a3ee2aa3bfeb8db] [formerly 023ded895139af797fbcc13d539515f225f28986 [formerly 7b5d8b672b3beb1e5c624ff7c2e8544096db117e]]
Former-commit-id: 6e5d81007a3eeb857400f66d4f8ab95eff02aa44 [formerly b66b408d86dd24aa3a0bc2398e58dfcac286fe2f]
Former-commit-id: 478fb27899a350b5d77e4cd84a1cf422f398396b