Merged commit r4164 by efocht from 2006-01-25 14:04:02 +0100 (Wed, 25 Jan 2006) from repository https://svn.oscar.openclustergroup.org/svn/oscar/tmp/ef_sandbox to repository https://svn.oscar.openclustergroup.org/svn/oscar/trunk Original commit message was:
Had to split this out of Package because it is needed during the prereqs
install which happens such early that we cannot be sure that XML::Simple
(required by OSCAR::Package) is already available. If all goes well, this
could replace the simple RPM_POOL variable provided by OSCAR::Package.

The package provides functions for building repository paths depending on
distro, version, etc... Currently only two functions are provided:

distro_repo_url($distro,$distro_version,$architecture)
------------------------------------------------------
The URL where distribution packages are stored. When the files are
stored locally they should go to:
/tftpboot/distro/$distro-$version-$arch
Doing this without subtrees of dirs has the advantage to immediately show
the admin how many of these disk space consuming distro repositories are
around.

If the nodes have connectivity to the internet one could use a publicly
accessible URL for the distro files. In that case place the URL of the
repository (yum or apt) into the file
/tftpboot/distro/$distro-$version-$arch.url


oscar_repo_path($distro,$distro_version,$architecture)
------------------------------------------------------
The path where OSCAR packages for a particular distro/version/arch
combination are stored. This path is defined as:
/tftpboot/oscar/$distro-$version-$arch

The distro and version names used are those detected as "compat_distro"
and "compat_distrover" in the OCA::OS_Detect framework. The reason is that
OSCAR doesn't care about the particular flavor of a rebuilt distro, it
uses the same packages eg. for rhel4, scientific linux 4 and centos 4.




git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@4247 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: e1cebe9850b5ecb064fc49a4fe5eac3b4010a36f [formerly 32f96279972a2b8313c97afa511d69747799feb9] [formerly 97227a7aef7095df09570a20638c697b6f34a228 [formerly 756a9be4393044ffd0b76058b92689bfd1ca9c15]] [formerly 25e919b71306c7ffdee7c36dd9aa4cc835755e60 [formerly 95466bd9a2639348909e2b7d3fc34acce02d1f73] [formerly ada0ba977ccf13e1097bd0b49f080debd43d3729 [formerly f15ec8a9dcb5e81cfac118428637bb9cebc20285]]]
Former-commit-id: 05659ee904b97fb9655ef84687574ebe8cf4d6f8 [formerly 6d70819ab16fe4dd3589f36c0e89ec4cac49976b] [formerly 9de3ff3087bceaa87faa86fd548d94a179b91de9 [formerly 893619aedcb475456c291f111fcdb6cb0ecf6a5b]]
Former-commit-id: 20d183a6766cb990321f408e9329b2682e6a884d [formerly 6741d736618e54ac48addc834cd4bf1a98b8e8a4]
Former-commit-id: 51996a8751e1ce53ae6f18a4449bdbd358d2173b