This is the rework of the build infrastructure I talked about at the phone conference.
Summary:
Instead of the sequence:
	./autogen.sh
	./configure
	make dist
you now just need to do
        make dist
The process is 10-20 times faster, depending on what you do.

automake and configure are gone. Also all Makefile.am are gone.

In order to ease testing and development one can now install directly the SVN
checkout by doing "make test".



If you want to test and develop directly on the OSCAR SVN do (inside the trunk
checkout):
   make test
This will set up the needed repositories, build the perl-Qt related files in
src but _not_ insert the copyright mesages and not build the docs. After this
you will be able to do:
   ./install_cluster eth0
from inside the trunk checkout.

If you finished your work you should try
   make clean
which will clean up the src directory and remove the checksum directory tmp/
such that your checkout will be as fresh as before.


If you want to build the OSCAR distribution you will need to do:
   make dist

This command will create a set of tarballs in the _parent_ directory of the
trunk checkout!

The OSCAR distibution comes in two OSCAR tarballs and a set of repository
tarballs:
  oscar-base-OSCARVERSION.tar.gz  : the base OSCAR tarball stripped of
                                    almost every binary packages (rpms)
  oscar-srpms-OSCARVERSION.tar.gz : the SRPMS, you only need this if you
                                    want to rebuild rpms
  oscar-repo-common-rpms-OSCARVERSION.tar.gz :
                                    repository of noarch rpms used on all
                                    supported distributions
  oscar-repo-DISTRO-VER-ARCH-OSCARVERSION.tar.gz :
                                    repository of distro and architecture
                                    specific rpms.

For an OSCAR installation:
 - Download the oscar-repo-common-rpms and the oscar-repo-DISTRO-VER-ARCH
 tarballls suitable for the distribution(s). Unpack them in /tftpboot/oscar.
 - Download the oscar-base tarball. Unpack it in an appropriate directory.
 Rename the unpacked directory, if you want (e.g. /opt/oscar)
 - In the unpacked directory (e.g. /opt/oscar) execute:
 ./install_cluster ETH_INTERFACE

 (ETH_INTERFACE is the cluster internal interface on the head node,
 e.g. eth0).




git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@4831 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 9c25cfadbbd6285d7cdfb0c72fc3175f3ab01659 [formerly c28aff3430422aa9cb4a1fede215e909dbc50892] [formerly 9a98866efc5070d11ce1ab53f4d23b31575a7e72 [formerly 5f181d10a1c5a983a7793df345bdcd1d0de8bf1e]] [formerly 2921e695303caff8cb1c2369625589e7dff87bcf [formerly bd1134c6fefc8ea1cf8a899f46400d3f166cb941] [formerly 4e74d622c4dec6e6e25e5c7b7711866e79ed4c67 [formerly 64f2dc0cbf1b873741c72a8d72df8d21ff83966a]]]
Former-commit-id: 5aa0bd4b49e367cd2cd4fcdc95b192a70d8560b2 [formerly 68f7eb5679a7e02277d3cff8dfa69cf2d196d45d] [formerly 76712360937c7c9563a37f0d7f87e369cf302c2f [formerly b20f04824159754b77ff83fbc09c35af67de9937]]
Former-commit-id: f0ae8c0e997a1954e1b1ca335798aee6785b54d5 [formerly a45a9566606e9908d87ac08be3d8c882d8e6c181]
Former-commit-id: 2697979b72cd724414fd094e8a38fbf5a4876314