Moved per opkg configuration variables to the OSCAR database. This means: .configurator.values files are not created any more! The values go to the Packages_config table in the OSCAR database.
The Configbox.pm module was modified a bit to access the configurator
variables from the ODA. Also some packages' API scripts needed changes. I
tested this to work for ganglia and ntpconfig, I hope I cought all scripts
which needed to be changed. If you find anything which still needs to be
changed, please tell me.

Configbox.pm takes care of the configurator values handling. Use
   use OSCAR::Configbox;
   my $values = readInConfigValues($cfile,$opkg,$context,[noarray=>1]);

The noarray arg is optional.
You'll get back a hash reference, as you were used to.

If you want to interact directly with the database use the routines:
  set_pkgconfig_var
  get_pkgconfig_vars
  del_pkgconfig_vars
  pkgconfig_values
from OSCAR::Database.pm (this is addressed to Wesley for CLI).

Otherwise nothing changed.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@5066 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 17013983092ae6987b0940074bc9f0984770a6be [formerly 3716d1257873b86bcb276bed606c122774361aa3] [formerly 3bec6a4c2fcc12f0e4f92171f9b2fdc3998688df [formerly e085a5c07ea7a7c02e1bf82eea2d17a8ab5a96b2]] [formerly 4f9c742243e444ca23b684fa369818d2a1073b41 [formerly 5a01288a6dc1704cb17afbcf0c52f6fdc0478aab] [formerly 8ca98909de1cf66d6ec21a7ea5ade5b2b9960a9a [formerly 827cded4d4a0e17f94c2ba534f509f00d79d45ce]]]
Former-commit-id: 5d6493c498cef9854a63b6910d43cd0120c42aea [formerly 5006887248c4102679c770aa8635347041519726] [formerly 9e0c4ea790af38966638bf039e4b2752caf1f784 [formerly e23b6b4a7aad431cc5d296c1c6b472fe7f732e07]]
Former-commit-id: 9a717fb32e3a37f90f9ed92adbc1b0fa58df13bd [formerly 6fc0b96a25e684ba2d0856371ca00974d3d778da]
Former-commit-id: 43a6318a71a63206e29ebf5b337e927003d85c4f