New schema for opkgc based OSCAR package management.
It makes little sense to carry the whole unnecessary garbage around.
Deleting the unneeded tables will immediately point us to the places
which need to be fixed.

The core opkg related tables are:
 Packages
 Node_Package_Status

Packages should actually be renamed to Opkgs.
It will contain the list of all opkgs visible in all accessible repositories.

The more I think about it, the less I believe Packages is needed at all.
We could populate it with the output of OpkgDB::opkg_list_available, but
those are only the packages available for the distribution of the master
node. That's not good!


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@6381 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: e42031301407070e83eb5d8dacf0aa957a50d1b4 [formerly 05c1fc3d2829f31399e9120ebf68422422850be1] [formerly a7cb10c3e1bc57c3358018e126cd662b77bea120 [formerly 00a5eb484503e6410053be8e4cacdafcfe14d6fb]] [formerly a6bb7d14fa9d41ba1c04b9117b5bd2e6ee8aabe7 [formerly 0293eb259bcda2f1b972497f9804a9697ae2d18e] [formerly 55df93da103c79790309b823e492a1db7a03e869 [formerly 500143d8a19d3413e1e0999fabe3d04063af7006]]]
Former-commit-id: 414f546dd339095624f1f8e99b55d72e33a2d01a [formerly 4e8652bd1ee810dab5e74377834e37a94a080b60] [formerly c8a433ad0106506b6c0eb578e8f55242ed39d390 [formerly bf9da453db1484e545e6f42fbd8630b0cf058674]]
Former-commit-id: f48324532ee6988873b71f9505041105a34db72b [formerly 5d3a8b7ef5cda248bdddc12030b43cc2a5c00292]
Former-commit-id: 2d2f6f75e010b287ce2950baecc5f7cbe33742fc